requirements.txt
pyTelegramBotAPI==4.15.4
gspread==6.1.2
google-auth==2.35.0